Spinning ant
The ant can now spin without everything going all bad. I initially tried continuously rotating the ant's image, but because of how the rotation was resampled, it made the ant turn into a big bunch of black pixels. I've now kept the original image, so rotations are based on that, with a `self` value for the direction (0-360 anticlockwise).