Update hole_class.py
For some reason, when trying to get the ants to appear at the coordinates where the holes are with the previous method,ants seem to only come from the [900,600] coordinate that is set by default within the hole_locations list. 
I think the problem might be related to the way in which values are located, when I tried to implement this method with a way of making the holes appear at the given coordinates other than the use of the choice function, the main issue was that only the first hole in the list appeared, whereas the rest of the ants randomly started at the coordinates generated by the list. 
However, when I set the range function within the main method that calls holes to say 20 and the number within holes_manager function to 5, just by probability the amount of holes that is needed appears, and ants randomly start from the coordinates generated by the list which also are the coordinates where the holes are. Everything seems to be working except for the holes clashing bug that was solved with the previous method. 
It looks as if one option inevitably denied the other, either we try to fix the holes clashing bug using another method while maintaining the code I designed for the ants to appear at the desired points, or we try to get the ants to come from the holes using the previous method which seems unlikely to be doable from my perspective, since when made the same modification I made in the ants_class without altering the holes_class in any way, ants only appeared at that point that is stored in the holes_location list by default. 
My personal view is, that despite losing the flexibility we would have by enabling the program to generate as many non-clashing holes as it can, only a maximum number of 10 holes I think would be created when running the program if we want it to look interesting, also the possibility of the holes clashing with such number of holes is considerably reduced with this code and after all it works. 
If it is possible to implement a better algorithm to have both features in the code that would obviously be all the better, but I am afraid that might not be the case due to the nature of the problem itself.