Ants appear from holes
Ants sometimes appear at random times from a random hole. When there are 0 holes, no ants can be initialised, so a message is displayed prompting to add a hole using the 'W' key, although this hasn't yet been implemented. This is related to issue #3, but I'm not closing it as ants don't get go into holes.